requirements.txt
Flask==2.3.2
Flask-Cors==3.1.1
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9   # If you use PostgreSQL
mysqlclient==2.1.1       # If you use MySQL
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0          # Recommended for Render deployment